+++ outbound/client_pool.js
@@ -7,13 +7,15 @@ const logger       = require('../logger');
 
 const obc          = require('./config');
 
+exports.name = 'outbound'
+
 function _create_socket (name, port, host, local_addr, is_unix_socket, callback) {
 
     const socket = is_unix_socket ? sock.connect({path: host}) : sock.connect({port, host, localAddress: local_addr});
     socket.name = name;
     socket.__uuid = utils.uuid();
     socket.setTimeout(obc.cfg.connect_timeout * 1000);
-    logger.logdebug(`[outbound] created. host: ${host} port: ${port}`, { uuid: socket.__uuid });
+    logger.logdebug(exports, `created. host: ${host} port: ${port}`, { uuid: socket.__uuid });
     socket.once('connect', () => {
         socket.removeAllListeners('error'); // these get added after callback
         socket.removeAllListeners('timeout');
@@ -23,26 +25,26 @@ function _create_socket (name, port, host, local_addr, is_unix_socket, callback)
         socket.end();
         socket.removeAllListeners();
         socket.destroy();
-        callback(`Outbound connection error: ${err}`, null);
+        callback(err.message, null);
     });
     socket.once('timeout', () => {
         socket.end();
         socket.removeAllListeners();
         socket.destroy();
-        callback(`Outbound connection timed out to ${host}:${port}`, null);
+        callback(`connection timed out to ${host}:${port}`, null);
     });
 }
 
-
 // Get a socket for the given attributes.
-exports.get_client = (port = 25, host = 'localhost', local_addr, is_unix_socket, callback) => {
-    const name = `outbound::${port}:${host}:${local_addr}`;
-
-    _create_socket(name, port, host, local_addr, is_unix_socket, callback)
+exports.get_client = function (port = 25, host = 'localhost', local_addr, is_unix_socket, callback) {
+    _create_socket(
+        `outbound::${port}:${host}:${local_addr}`,
+        ...arguments
+    )
 }
 
 exports.release_client = (socket, port, host, local_addr, error) => {
-    logger.logdebug(`[outbound] release_client: ${socket.__uuid} ${host}:${port} to ${local_addr}`);
+    logger.logdebug(exports, `release_client: ${socket.__uuid} ${host}:${port} to ${local_addr}`);
     socket.removeAllListeners();
     socket.destroy();
 }