+++ outbound/config.js
@@ -3,6 +3,8 @@
 const config = require('haraka-config');
 const logger = require('../logger');
 
+exports.name = 'outbound/config'
+
 function load_config () {
     const cfg = exports.cfg = config.get('outbound.ini', {
         booleans: [
@@ -64,8 +66,8 @@ exports.set_temp_fail_intervals = function () {
 
     // Helpful error function in case of parsing failure
     function error (i, msg) {
-        logger.logerror(`outbound temp_fail_intervals syntax error parsing element ${i}: ${msg}`);
-        logger.logwarn('Setting outbound temp_fail_intervals to old defaults because of previous error');
+        logger.logerror(exports, `temp_fail_intervals syntax error parsing element ${i}: ${msg}`);
+        logger.logwarn(exports, 'Setting outbound temp_fail_intervals to old defaults');
         set_old_defaults();
     }